Removed unnecessary uint64 castings
After we updated block height to uint64 we no longer need to cast it in
couple of places. This commit removes not needed castings.

+++ pkg/chain/local/local_test.go
@@ -199,7 +199,7 @@ func TestLocalOnRelayEntryGenerated(t *testing.T) {
 		Value:       big.NewInt(19),
 		GroupPubKey: []byte("1"),
 		Seed:        big.NewInt(30),
-		BlockNumber: uint64(123),
+		BlockNumber: 123,
 	}
 
 	chainHandle.SubmitRelayEntry(expectedEntry)
